string: fixed compiler warning
git-svn-id: http://svn.ta3d.org/yuni/trunk@2462 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/string/string.hxx
@@ -3490,7 +3490,7 @@ namespace Yuni
 	{
 		size_t hash = 0;
 		for (uint i = 0; i != AncestorType::size; ++i)
-			hash = AncestorType::data[i] + (hash << 6) + (hash << 16) - hash;
+			hash = (uint)AncestorType::data[i] + (hash << 6) + (hash << 16) - hash;
 
 		return hash;
 	}
